apply changes from code review, fix many compiler warnings
