+++ DatFileRenamer/Program.cs
@@ -1,6 +1,4 @@
-﻿using System.Diagnostics;
-using System.IO;
-using System.Reflection;
+﻿using System.Reflection;
 
 Console.WriteLine("Dat File Renamer v0.1");
