+++ OpenLocoTool/DatFileParsing/AttributeHelper.cs
@@ -12,12 +12,11 @@ namespace OpenLocoTool.DatFileParsing
 
 		public static T? Get<T>(Type t) where T : Attribute
 			=> t.GetCustomAttribute(typeof(T), inherit: false) as T;
-		public static IEnumerable<T?> GetAll<T>(Type t) where T : Attribute
-			=> t.GetProperties().Select(p => p.GetCustomAttribute(typeof(T), inherit: false) as T);
+
 		public static bool Has<T>(PropertyInfo p) where T : Attribute
 			=> p.GetCustomAttribute(typeof(T), inherit: false) is T;
 
-		public static bool Has<T>(Type t) where T : Attribute
-			=> t.GetCustomAttribute(typeof(T), inherit: false) is T;
+		public static IEnumerable<PropertyInfo> GetAllPropertiesWithAttribute<T>(Type t) where T : Attribute
+			=> t.GetProperties().Where(Has<T>);
 	}
 }
