+++ OpenLocoTool/DatFileParsing/ILocoObject.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel;
+using OpenLocoTool.Headers;
+
+namespace OpenLocoTool.DatFileParsing
+{
+	[TypeConverter(typeof(ExpandableObjectConverter))]
+	public interface ILocoObject
+	{
+		S5Header S5Header { get; set; }
+		ObjectHeader ObjectHeader { get; set; }
+		ILocoStruct Object { get; set; }
+		StringTable StringTable { get; set; }
+		G1Header G1Header { get; set; }
+		List<G1Element32> G1Elements { get; set; }
+	}
+}
