+++ OpenLocoTool/DatFileParsing/ILocoStructStringTablePostLoad.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel;
+
+namespace OpenLocoTool.DatFileParsing
+{
+	[TypeConverter(typeof(ExpandableObjectConverter))]
+	public interface ILocoStructStringTablePostLoad
+	{
+		void LoadPostStringTable(StringTable stringTable);
+	}
+}
