+++ OpenLocoTool/DatFileParsing/ILocoStructVariableData.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+
+namespace OpenLocoTool.DatFileParsing
+{
+	[TypeConverter(typeof(ExpandableObjectConverter))]
+	public interface ILocoStructVariableData
+	{
+		ReadOnlySpan<byte> Load(ReadOnlySpan<byte> remainingData);
+
+		ReadOnlySpan<byte> Save();
+	}
+}
