+++ OpenLocoTool/DatFileParsing/LocoAttributes.cs
@@ -18,14 +18,6 @@
 		public int Size => size;
 	}
 
-	//[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = false)]
-	//public class LocoStringTableAttribute(params string[] names) : Attribute
-	//{
-	//	public string[] Names => names;
-
-	//	public int Count => Names.Length;
-	//}
-
 	// basically a 'skip' attribute to allow deferred loading for variable data
 	[AttributeUsage(AttributeTargets.Parameter | AttributeTargets.Property, AllowMultiple = false)]
 	public class LocoStructVariableLoadAttribute : Attribute
